fe: Rename SortButton component function to SortDropdown (yes, I renamed the file but forgot to rename the function the first time)
